Switch to libwes64 VECTOR
Also switch entire program from double to float

+++ src/color.c
@@ -11,9 +11,9 @@
  * Convert RGB to COLOR
  *============================================================*/
 void rgb_ToColor(COLOR *color, const RGB *rgb) {
-    color->x = (double)rgb->r / RGB_MAX;
-    color->y = (double)rgb->g / RGB_MAX;
-    color->z = (double)rgb->b / RGB_MAX;
+    color->x = (float)rgb->r / RGB_MAX;
+    color->y = (float)rgb->g / RGB_MAX;
+    color->z = (float)rgb->b / RGB_MAX;
 }
 
 void color_ToRgb(RGB *rgb, const COLOR *color) {
